2-langchain-agent/requirements.txt
asana==5.0.10
langchain_anthropic==0.3.0
langchain_core==0.3.21
langchain_openai==0.2.11
openai==1.57.0
python-dotenv==1.0.1
